refactor: GamePage hooks Barrel Parttern 적용 #439
