+++ frontend/src/assets/index.ts
@@ -1,0 +1,39 @@
+import AngryDdangkong from './images/angryDdangkong.webp';
+import ArrowDown from './images/arrowDown.svg';
+import ArrowLeft from './images/arrowLeft.svg';
+import ArrowUp from './images/arrowUp.svg';
+import CloseIcon from './images/closeIcon.png';
+import CopyIcon from './images/copyIcon.png';
+import CrownIcon from './images/crownIcon.webp';
+import Ddangkong from './images/ddangkong.webp';
+import DdangkongTimer from './images/ddangkongTimer.webp';
+import ErrorDdangkong from './images/errorDdangkong.webp';
+import ExitIcon from './images/exitIcon.svg';
+import HomeIcon from './images/homeIcon.svg';
+import SadDdangkong from './images/sadDdangkong.webp';
+import SettingIcon from './images/settingIcon.svg';
+import SillyDdangkong from './images/sillyDdangkong.webp';
+import SillyDdangkongMedium from './images/sillyDdangkongMedium.webp';
+import SillyDdangkongSmall from './images/sillyDdangkongSmall.webp';
+import SpinDdangkong from './images/spinDdangkong.webp';
+
+export {
+  AngryDdangkong,
+  ArrowDown,
+  ArrowLeft,
+  ArrowUp,
+  CloseIcon,
+  CopyIcon,
+  CrownIcon,
+  Ddangkong,
+  DdangkongTimer,
+  ErrorDdangkong,
+  ExitIcon,
+  HomeIcon,
+  SadDdangkong,
+  SettingIcon,
+  SillyDdangkong,
+  SillyDdangkongMedium,
+  SillyDdangkongSmall,
+  SpinDdangkong,
+};
