+++ frontend/src/pages/GamePage/components/SelectContainer/SelectContainer.tsx
@@ -2,10 +2,10 @@ import { useParams } from 'react-router-dom';
 
 import GameVoteStatusContainer from './GameVoteStatusContainer/GameVoteStatusContainer';
 import useSelectOption from './hooks/useSelectOption';
+import SelectButton from './SelectButton/SelectButton';
 import { selectContainerLayout, selectSection } from './SelectContainer.styled';
 import SelectOption from './SelectOption/SelectOption';
 import Timer from './Timer/Timer';
-import SelectButton from '../SelectButton/SelectButton';
 
 import useBalanceContentQuery from '@/hooks/useBalanceContentQuery';
