+++ frontend/src/pages/GameResultPage/components/GameResult/hooks/index.ts
@@ -1,0 +1,5 @@
+import useMatchingResultQuery from './useMatchingResultQuery';
+import useScrollControl from './useScrollControl';
+import useScrollState from './useScrollState';
+
+export { useMatchingResultQuery, useScrollState, useScrollControl };
